checked in

+++ pisicasrv/skimg.cpp
@@ -41,7 +41,7 @@ int skimg::snd(const uint8_t* b, size_t len,uint32_t extra)
                                     (int)timestamp.tv_sec,
                                     (int)timestamp.tv_usec);
     int rv = skweb::snd((const uint8_t*)buffer,hl);
-    if(rv == hl)
+    if(rv == (int)hl)
     {
         skweb::snd(b,len);
         skweb::snd((const uint8_t*)JpegPart,strlen(JpegPart));
